Speedup of level reloading in OpenGL mode
Most of level's structures now are allocated with PU_STATIC instead of PU_LEVEL. It is important for OpenGL, because in case of the same data in memory we can skip recalculation of much stuff.

Loadgame from the current (same) level now is much faster (100x). You can try it with Epic.wad\map05 (save, load, load...)

git-svn-id: https://svn.prboom.org/repos/branches/prboom-plus-24/prboom2@2733 df9a1a3e-e9e3-0310-a482-b5a01ba5a74e