One more 'ATI sucks' revision.
Some ATI’s drivers have a bug whereby adding the depth renderbuffer and then a texture causes the application to crash. This should be kept in mind when doing any FBO related work and tested for as it is possible it could be fixed in a future driver revision thus rendering the problem non-existent.

Some TRY/EXCEPT stuff and workarounds here

git-svn-id: https://svn.prboom.org/repos/branches/prboom-plus-24/prboom2@2796 df9a1a3e-e9e3-0310-a482-b5a01ba5a74e