m_misc.[ch]: macro version of M_DoubleToInt
On GCC, use an extension (compound statement expressions[1]) to declare
a temporary variable; on others, just cast the parameter to double.

I'm not sure why casting the return value of a function from double to
int causes a warning, but putting the result in a temporary double and
casting that to int, does not.

[1] http://gcc.gnu.org/onlinedocs/gcc/Statement-Exprs.html

git-svn-id: https://svn.prboom.org/repos/branches/prboom-plus-24/prboom2@3948 df9a1a3e-e9e3-0310-a482-b5a01ba5a74e