r_patch.c: FillEmptySpace: copy top row to bottom
Copy any solid pixels across the top row of the patch into any
corresponding space at the bottom of the same column.

This is a hack which fixes the erroneously-coloured row of pixels
across the top of the firing chaingun weapon sprite.

{The real problem is rounding error in R_DrawColumn. For sprites and
masked middle textures we would really like frac to be clamped to
[0,texheight) instead of wrapping to the other end of the interval.}

git-svn-id: https://svn.prboom.org/repos/branches/prboom-plus-24/prboom2@4431 df9a1a3e-e9e3-0310-a482-b5a01ba5a74e