playpal_transparent: find duplicate of palette entry zero
Find a palette entry that is the duplicate of palette entry zero,
so we can use zero as a transparency index in FillEmptySpace.

This allows patches to use palette entry 255 without corruption [1].
Note: ZDoom assumes palette entry zero will have a duplicate, and even
wads with highly de-duplicated palettes (e.g. BTSX) seem to maintain it.
Thus I believe it to be a convention sufficiently widespread to adopt.

[1] http://www.doomworld.com/vb/post/1381944

git-svn-id: https://svn.prboom.org/repos/branches/prboom-plus-24/prboom2@4446 df9a1a3e-e9e3-0310-a482-b5a01ba5a74e